relocate code into a module
