+++ packages/dappmanager/src/calls/stakerConfig.ts
@@ -1,380 +1,25 @@
-import * as db from "../db";
-import { logs } from "../logs";
-import { ComposeFileEditor } from "../modules/compose/editor";
-import {
-  Network,
-  PkgStatus,
-  StakerConfigGet,
-  StakerConfigSet,
-  UserSettingsAllDnps
-} from "../types";
-import { packageInstall } from "./packageInstall";
-import { packageSetEnvironment } from "./packageSetEnvironment";
-import { packagesGet } from "./packagesGet";
-import { packageStartStop } from "./packageStartStop";
+import { getStakerConfig } from "../modules/stakerConfig/getStakerConfig";
+import { setStakerConfig } from "../modules/stakerConfig/setStakerConfig";
+import { Network, StakerConfigGet, StakerConfigSet } from "../types";
 
 /**
- * Sets a new staker configuration based on user selection:
- * - New execution client
- * - New consensus client
- * - Install web3signer and/or mevboost
- * - graffiti and fee recipient address
- * @param stakerConfig
- * TODO: add option to remove previous or not
+ * Sets the staker configuration: execution and consensus clients, remote signer,
+ * mev boost, graffiti, fee recipient address and checkpoint sync url
  */
 export async function stakerConfigSet({
   stakerConfig
 }: {
   stakerConfig: StakerConfigSet;
 }): Promise<void> {
-  const {
-    execClientsAvail,
-    currentExecClient,
-    consClientsAvail,
-    currentConsClient,
-    web3signerAvail,
-    mevBoostAvail
-  } = getNetworkStakerPkgs(stakerConfig.network);
-
-  try {
-    // Ensure Execution clients DNP's names are valid
-    if (
-      stakerConfig.executionClient &&
-      !execClientsAvail.includes(stakerConfig.executionClient)
-    )
-      throw Error(
-        `Invalid execution client ${stakerConfig.executionClient} for network ${stakerConfig.network}`
-      );
-    // Ensure Consensus clients DNP's names are valid
-    if (
-      stakerConfig.consensusClient &&
-      !consClientsAvail.includes(stakerConfig.consensusClient)
-    )
-      throw Error(
-        `Invalid consensus client ${stakerConfig.consensusClient} for network ${stakerConfig.network}`
-      );
-
-    const pkgs = await packagesGet();
-
-    // EXECUTION CLIENT
-    if (stakerConfig.executionClient) {
-      if (currentExecClient === stakerConfig.executionClient) {
-        logs.info(
-          "Execution client is already set to " + stakerConfig.executionClient
-        );
-        // Make sure the EC is installed
-        if (!pkgs.find(p => p.dnpName === stakerConfig.executionClient)) {
-          logs.info("Installing " + stakerConfig.executionClient);
-          await packageInstall({ name: stakerConfig.executionClient });
-        }
-      } else {
-        // Remove the previous
-        /* logs.info("Removing " + currentExecClient);
-          await packageRemove({ dnpName: currentExecClient }); */
-        // Install the new EC
-        logs.info("Installing " + stakerConfig.executionClient);
-        await packageInstall({ name: stakerConfig.executionClient });
-      }
-    }
-
-    // CONSENSUS CLIENT (+ Fee recipient address + Graffiti + Checkpointsync)
-    if (stakerConfig.consensusClient) {
-      const userSettings: UserSettingsAllDnps = {
-        [stakerConfig.consensusClient]: {
-          environment: {
-            [getValidatorServiceName(stakerConfig.consensusClient)]: {
-              // Graffiti is a mandatory value
-              ["GRAFFITI"]: stakerConfig.graffiti || "Validating_from_DAppNode",
-              // Fee recipient is a mandatory value
-              ["FEE_RECIPIENT_ADDRESS"]:
-                stakerConfig.feeRecipient ||
-                "0x0000000000000000000000000000000000000000",
-              // Checkpoint sync is an optional value
-              ["CHECKPOINT_SYNC_URL"]: stakerConfig.checkpointSync || ""
-            }
-          }
-        }
-      };
-
-      if (currentConsClient === stakerConfig.consensusClient) {
-        logs.info(
-          "Consensus client is already set to " + stakerConfig.consensusClient
-        );
-        // Make sure the CC is installed
-        if (!pkgs.find(p => p.dnpName === stakerConfig.consensusClient)) {
-          logs.info("Installing " + stakerConfig.consensusClient);
-          await packageInstall({
-            name: stakerConfig.consensusClient,
-            userSettings
-          });
-          // Make sure to set the new environment (if any)
-        } else if (
-          stakerConfig.graffiti ||
-          stakerConfig.feeRecipient ||
-          stakerConfig.checkpointSync
-        ) {
-          const serviceEnv =
-            userSettings[stakerConfig.consensusClient].environment;
-
-          if (serviceEnv) {
-            logs.info(
-              "Updating environment for " + stakerConfig.consensusClient
-            );
-            await packageSetEnvironment({
-              dnpName: stakerConfig.consensusClient,
-              environmentByService: serviceEnv
-            });
-          }
-        }
-      } else {
-        // Remove the previous
-        /*       logs.info("Removing " + currentConsClient);
-        await packageRemove({ dnpName: currentConsClient }); */
-        // Install the new CC
-        logs.info("Installing " + stakerConfig.consensusClient);
-        await packageInstall({
-          name: stakerConfig.consensusClient,
-          userSettings
-        });
-      }
-    }
-
-    // WEB3SIGNER
-    const web3signerPkg = pkgs.find(p => p.dnpName === web3signerAvail);
-    if (web3signerPkg && stakerConfig.enableWeb3signer) {
-      // Do nothing, web3signer enabled and installed
-      logs.info("Web3Signer is already installed");
-    } else if (!web3signerPkg && stakerConfig.enableWeb3signer) {
-      // Install web3signer
-      logs.info("Installing Web3Signer");
-      await packageInstall({ name: web3signerAvail });
-    } else if (web3signerPkg && !stakerConfig.enableWeb3signer) {
-      // Stop web3signer
-      for (const container of web3signerPkg.containers) {
-        if (container.running) {
-          logs.info("Stopping Web3Signer");
-          await packageStartStop({
-            dnpName: web3signerPkg.dnpName,
-            serviceNames: [container.serviceName]
-          }).catch(e => logs.error(e.message));
-        }
-      }
-    }
-
-    // MEV BOOST
-    const mevBoostPkg = pkgs.find(p => p.dnpName === mevBoostAvail);
-    if (stakerConfig.enableMevBoost) {
-      if (mevBoostPkg) {
-        logs.info("MevBoost is already installed");
-      } else if (stakerConfig.enableMevBoost) {
-        // Install mevboost if selected and not installed
-        logs.info("Installing MevBoost");
-        await packageInstall({ name: mevBoostAvail });
-      }
-    }
-
-    // Persist the staker config on db
-    setStakerConfigOnDb(stakerConfig.network, stakerConfig);
-  } catch (e) {
-    throw Error(`Error setting staker config: ${e}`);
-  }
+  await setStakerConfig({ stakerConfig });
 }
 
 /**
- * Fetches the current staker configuration:
- * - execution clients: isInstalled and isSelected
- * - consensus clients: isInstalled and isSelected
- * - web3signer: isInstalled and isSelected
- * - mevBoost: isInstalled and isSelected
- * - graffiti
- * - fee recipient address
- * - checkpoint sync url
- * @param network
+ * Returns the current staker configuration: execution and consensus clients,
+ * remote signer, mev boost, graffiti, fee recipient address and checkpoint sync url
  */
 export async function stakerConfigGet(
   network: Network
 ): Promise<StakerConfigGet> {
-  try {
-    const {
-      execClientsAvail,
-      currentExecClient,
-      consClientsAvail,
-      currentConsClient,
-      web3signerAvail,
-      mevBoostAvail,
-      isMevBoostSelected
-    } = getNetworkStakerPkgs(network);
-
-    const pkgs = await packagesGet();
-
-    // Execution clients
-    const executionClients: PkgStatus[] = [];
-    for (const exCl of execClientsAvail) {
-      executionClients.push({
-        dnpName: exCl,
-        isInstalled: pkgs.some(pkg => pkg.dnpName === exCl),
-        isSelected: currentExecClient === exCl
-      });
-    }
-
-    // Consensus clients
-    const consensusClients: PkgStatus[] = [];
-    for (const conCl of consClientsAvail) {
-      consensusClients.push({
-        dnpName: conCl,
-        isInstalled: pkgs.some(pkg => pkg.dnpName === conCl),
-        isSelected: currentConsClient === conCl
-      });
-    }
-
-    // Web3signer
-    const isWeb3signer = pkgs.some(pkg => pkg.dnpName === web3signerAvail);
-    const web3signer = {
-      dnpName: web3signerAvail,
-      isInstalled: isWeb3signer,
-      isSelected: isWeb3signer
-    };
-
-    // Mevboost
-    const mevBoost = {
-      dnpName: mevBoostAvail,
-      isInstalled: pkgs.some(pkg => pkg.dnpName === mevBoostAvail),
-      isSelected: isMevBoostSelected
-    };
-
-    // Get graffiti, feerecipientAddress and checkpointSync from the consClientInstalled (if any)
-    const consensusPkgSelected = pkgs.find(
-      pkg => pkg.dnpName === currentConsClient
-    );
-    let graffiti = "";
-    let feeRecipient = "";
-    let checkpointSync = "";
-    if (consensusPkgSelected) {
-      const environment = new ComposeFileEditor(
-        consensusPkgSelected.dnpName,
-        consensusPkgSelected.isCore
-      ).getUserSettings().environment;
-      if (environment) {
-        const validatorService = getValidatorServiceName(
-          consensusPkgSelected.dnpName
-        );
-        graffiti = environment[validatorService]["GRAFFITI"];
-        feeRecipient = environment[validatorService]["FEE_RECIPIENT_ADDRESS"];
-        checkpointSync = environment[validatorService]["CHECKPOINT_SYNC_URL"];
-      }
-    }
-
-    return {
-      executionClients,
-      consensusClients,
-      web3signer,
-      mevBoost,
-      graffiti,
-      feeRecipient,
-      checkpointSync
-    };
-  } catch (e) {
-    throw Error(`Error getting staker config: ${e}`);
-  }
-}
-
-// Utils
-
-/**
- * Sets the staker configuration on db for a given network
- */
-function setStakerConfigOnDb(
-  network: Network,
-  stakerConfig: StakerConfigSet
-): void {
-  switch (network) {
-    case "mainnet":
-      db.executionClientMainnet.set(stakerConfig.executionClient);
-      db.consensusClientMainnet.set(stakerConfig.consensusClient);
-      db.mevBoostMainnet.set(stakerConfig.enableMevBoost);
-    case "gnosis":
-      db.executionClientGnosis.set(stakerConfig.executionClient);
-      db.consensusClientGnosis.set(stakerConfig.consensusClient);
-      db.mevBoostGnosis.set(stakerConfig.enableMevBoost);
-    case "prater":
-      db.executionClientPrater.set(stakerConfig.executionClient);
-      db.consensusClientPrater.set(stakerConfig.consensusClient);
-      db.mevBoostPrater.set(stakerConfig.enableMevBoost);
-  }
-}
-
-/**
- * Get the current staker config (execution and consensus clients selected) as well as
- * the pkgs available for each network
- */
-export function getNetworkStakerPkgs(network: Network): {
-  execClientsAvail: string[];
-  currentExecClient: string;
-  consClientsAvail: string[];
-  currentConsClient: string;
-  web3signerAvail: string;
-  mevBoostAvail: string;
-  isMevBoostSelected: boolean;
-} {
-  switch (network) {
-    case "mainnet":
-      return {
-        execClientsAvail: [
-          "geth.dnp.dappnode.eth",
-          "nethermind.dnp.dappnode.eth",
-          "besu.dnp.dappnode.eth",
-          "erigon.dnp.dappnode.eth"
-        ],
-        currentExecClient: db.executionClientMainnet.get(),
-        consClientsAvail: [
-          "prysm.dnp.dappnode.eth",
-          "lighthouse.dnp.dappnode.eth",
-          "teku.dnp.dappnode.eth",
-          "nimbus.dnp.dappnode.eth"
-        ],
-        currentConsClient: db.consensusClientMainnet.get(),
-        web3signerAvail: "web3signer.dnp.dappnode.eth",
-        mevBoostAvail: "mevboost.dnp.dappnode.eth",
-        isMevBoostSelected: db.mevBoostMainnet.get()
-      };
-
-    case "gnosis":
-      return {
-        execClientsAvail: ["nethermind-xdai.dnp.dappnode.eth"],
-        currentExecClient: db.executionClientGnosis.get(),
-        consClientsAvail: [
-          "gnosis-beacon-chain-prysm.dnp.dappnode.eth",
-          "lighthouse-gnosis.dnp.dappnode.eth",
-          "teku-gnosis.dnp.dappnode.eth",
-          "nimbus-gnosis.dnp.dappnode.eth"
-        ],
-        currentConsClient: db.consensusClientGnosis.get(),
-        web3signerAvail: "web3signer-gnosis.dnp.dappnode.eth",
-        mevBoostAvail: "mevboost-gnosis.dnp.dappnode.eth",
-        isMevBoostSelected: db.mevBoostGnosis.get()
-      };
-    case "prater":
-      return {
-        execClientsAvail: ["goerli-geth.dnp.dappnode.eth"],
-        currentExecClient: db.executionClientPrater.get(),
-        consClientsAvail: [
-          "prysm-prater.dnp.dappnode.eth",
-          "lighthouse-prater.dnp.dappnode.eth",
-          "teku-prater.dnp.dappnode.eth",
-          "nimbus-prater.dnp.dappnode.eth"
-        ],
-        currentConsClient: db.consensusClientPrater.get(),
-        web3signerAvail: "web3signer-prater.dnp.dappnode.eth",
-        mevBoostAvail: "mevboost-prater.dnp.dappnode.eth",
-        isMevBoostSelected: db.mevBoostPrater.get()
-      };
-  }
-}
-
-/**
- * Get the validator service name.
- * Nimbus package is monoservice (beacon-validator)
- */
-function getValidatorServiceName(dnpName: string): string {
-  return dnpName.includes("nimbus") ? "beacon-validator" : "validator";
+  return await getStakerConfig(network);
 }
